rdn2: remove incorrect configuration of system counter increment value
On the RD-N2 config-3 platform, the system counter per-tick increment
value is configured after the system counter has been enabled. This is
incorrect as the system counter has to count at 1GHz from the start when
enabled. So remove this incorrect configuration of the system counter
per-tick increment value register.

Now that the module 'syscnt_impdef' is enabled for all RD-N2 platform
and variants, that is sufficient to ensure that the per-tick increment
value is configured before the system counter is started.

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: Ic9cede5a1b972f3570200557ec5f65aee70d5138