neoverse-rd/interface: add a generic interface for address remapper
During the boot process, SCP requires access to the AP's address space
to set up devices or peripherals before the AP is turned on. SCP
address space is architecturally limited to a total size of 4GiB and
hence, it cannot map the entire AP address space.

To overcome this, the SCP subsystem incorporates an address translation
window (typically of size 1MiB) to remap or translate addresses in this
window region to the AP address space. The way this window regions is
configured varies between different Neoverse RD Platforms and, hence
its drivers.

To unify the way this address translation is configured, introduce an
interface for the platform address translation unit driver to provide
concrete implementations. Read and write interfaces are introduced, and
the driver which implements this interface is expected to configure the
address translation window to map the SCP address to the AP address
space.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: I171220132ecaf6aa5263513ed79f175c3e654310