module/cmn_cyprus: add support for HN-S target type in SCGs
The default target type for SCG regions is HN-F node. But when the Local
Coherency Node (LCN) feature is enabled, the SCG region target type must
be configured as HN-S. Hence, add a flag to enable LCN programming and
choose the target type for SCG based on the flag. While at it, introduce
other sam node types.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I5b81e7ccdc853c3c3c7b4e6fb54a3dda5ec41fd5