tools/check_api: modularize and improve readability
This patch simplifies the way that `check_api.py` script works by
creating helper functions and reducing the number of nested for
loops.
It also updates the `EXCLUDE_DIRECTORIES` variable to include
unit testing excluded directories. Lastly, it introduces the
possibility to select an external file for banned API check.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I333e6fb91d2729c9c49463c02ca4cb68a49ce88b