tools/check_build: Add check build to a separate file with YAML integration
This patch introduces a new script, check_build.py, to verify
the build process for various products. It modularizes the functionality
previously found in the ci_cmake.py test script, which executed a
predefined list of products with their respective configurations.

Furthermore, it removes all hardcoded product build definitions by
replacing them with a YAML configuration file. By default, it is set
to use the `product_build.yml` file.
Additionally, `*.yml` and `*.yaml` are added to the `FILE_TYPES`
in `check_spacing.py` script for consistency.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: Ib7df2dc5c2ba122d3154fcf0f48eb7ce75c4ab7a