rdfremont: reorganize common definitions
The refclk speed is a common definition for SCP, MCP and LCP. So remove
SCP and MCP specific definitions of refclk speed and consolidate it
under a common css definitions file. In addition to this, add the
per-tick system counter increment value definition that is required to
comply with the SBSA requirement of system counter incrementing at 1GHz.

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: Iebe89ddf1cd07468baefcfed432d58044e476487