power_domain: add new deeper state locking states
This patch includes the three new power domain states: OFF0,
OFF1 and OFF2 to the power domain module.

This will be used for the deeper locking states and will only
be passed over to a driver when it is being applied to a core.

This patch also includes some unit tests for the power domain
module to verify the new states being introduced, and how they
can be used when setting a state with a driver.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I89e541f2bf6e614e8588196dbb138a7ef104d4fd