mod_sensor: Add bool "disabled" for all sensors
This patch adds a boolean flag to the mod_sensor_info
structure. This flag is called "disabled".

If this flag is set to "true", then the sensor
starts in a disabled state.

If set to false, then the sensor starts in an
enabled state.

In addition, the patch adds initialisation of
"disabled" flag for all sensors on all platforms.

Together, these changes allow sensors to start in an
enabled or a disabled state.

Change-Id: Ia186621af4ca252cc08a7e2bb673a8df7693acab
Signed-off-by: Andrew Graham <andrew.graham@arm.com>