docker: Update Dockerfile and cleanup dependencies
This patch refactors the Dockerfile setup process, by optimizing
the installation of tools and dependencies. It simplifies
the environment setup by using ARGs for tool versions. This
updated Dockerfile allows to be used for `x86_64` and `aarch64`
host machines.
Also, it fixes the group ID extraction bug in the init script.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I38b999b2c54230af5e26b359823815f3194de511