module/cmn_cyprus: program AgentID-to-LinkID LUT
Remote agents, RAs or HAs, are identified using their RAID or HAID
respectively. Each remote RA or HA that a CML gateway can communicate
with must be behind only one link. So, assign a unique LinkID to each
remote CML protocol link by configuring the AgentID-to-LinkID LUT
registers in CCG RA and CCG HA nodes.

Note: The CCG nodes only support link id 0 and the reset value of the
AgentID-to-LinkID register is also 0. Hence, the link id programming is
skipped and the default values are set as valid.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I8c64532b99f0227bffc445db9a38d14cb9d27c6f