tools/filter_coverage_report: Consolidate all generate_coverage_report.py
Fwk and module have their own script to generate code coverage. Product
is also going to have support for unit test soon. This means another
similiar script needs to be created for product.

This patch allows all the directory to share a single script.
The script also allow any new entry to be added easily by adding the
filename and regex in the script.

The script filename is also renamed to filter_coverage_report to better
reflect it's functionality which is to filter out test and mock code
from the report.

Signed-off-by: Wen Ping Teh <wenping.teh@arm.com>
Change-Id: I5913188f12f3090d56490e3cb5ae24e56d4ded2d