scmi_power_domain: Add error handling for failed set_state
This patch addresses an issue that will handle returning an SCMI
if the set_state request has failed and the power domain is in
sync mode. If the state request passes the initial sanity checking,
the response will be handled correctly. However if this is not
the case, it would return to the SCMI PD and never respond.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: Ic72992b052bda107c8d0719059d38e8d0491b8f8