smcf: Check MODE_STAT after setting MODE_REQ
This patch introduces a check in the MGI module to ensure MODE_REQ
registers is set successfully. From the RTL design, MODE_REQ registers
should be configured individually and the register has to be processed
before setting the next MODE_REQ register. This is done by checking the
MODE_STAT register and a timeout is added to prevent indefinite wait.

Signed-off-by: Wen Ping Teh <wenping.teh@arm.com>
Change-Id: I6fe0469dea9a68f6c7374596b5566cdafeb17fcb