neoverse-rd/rdfremont: initialise event object to 0
The event object instance is defined within a function scope, and such
variables are allocated on the stack. In this case, only two fields are
initialised. Because the variable is allocated on the stack, other
fields could have incorrect values, leading to unexpected behaviour.

Zero initialise the event object to ensure that the unused fields are
set to a correct value.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: If8199a0ed693b18a927367f4501b339d721a5b8e