ci/platform-test: add platform boot tests
This patch adds a series of platform tests to be executed on FVP
or Board. It is included in the daily pipeline. Additionally,
it is added in the deployment pipeline, which is triggered
automatically for external public MRs or can be triggered
manually for internal MRs.
When `FETCH_PUBLIC_MR` is set to false, platform tests can be
triggered manually. Their status will not affect the overall
pipeline status due to limitations in GitLab CI, which does
not allow jobs to be both manual and optional while still
influencing the pipeline status.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: Ic3974a26037b12d3a2a862fa43dd945110591193