Add perf_controller module
This commit adds the skeleton of the  performance controller module.
It will replace the DVFS module in receiveing the performance set
requests from the SCMI perf.

The performance controller receives the power limit request by
implementing the generic set_limit interface. The power limit request could
limit the performance output if the power limit correpsonds to a lower performance
level that is lower than requested from the SCMI perf. This feature will be added
in the following commits.

The affect of introducing this module would be explained in the subsequent commits
that include the integration with SCMI perf, DVFS module and SCMI power capping.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>