mod/scmi_perf: integrate SCMI perf with perf_controller
The SCMI_perf module doesn't follow the same layered architecture as
the other protocols which use a HAL to communicate with different drivers.
Instead it is tightly couple to one driver which is the DVFS module.

With the introduction of the perf_controller module, it will act as the
HAL for the SCMI_perf. To limit the scope of the change as it will be
very large to untie the tight coupling between the SCMI_perf and the DVFS
module, a gradual integration of the performance controller module is done
where the performance controller module will only break the set level
link between the SCMI_perf and DVFS and receive the set level requests
from the SCMI_perf.

In this patch only the changes to the SCMI_perf are added for simplicity.
the changes to the DVFS module will be added in the next commits.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>