mod/dvfs: integrate with perf_controller
As the perf_controller breaks the tight coupling between the SCMI_perf
and DVFS, the DVFS will have some changes accordingly.

The DVFS will need to implement the driver interface for the
performance controller which for now will support the set_performance_level
request while still supporting the rest of the APIs required by the SCMI_perf.

This changes are implemented in this patch.

Signed-off-by: Ahmed Gadallah <ahmed.gadallah@arm.com>