mod/gtimer: return 0 if timestamp registers uninitalised
If a call is made to request a timestamp, which uses the element
configuration to read the corresponding registers, then it will
now return 0 if the module has yet to be initialised. This prevents
erroneously reading a memory location prior to it being set up
ensuring it fails in a way that does not crash the system.

Signed-off-by: Katherine Vincent <katherine.vincent@arm.com>
Change-Id: I3c896823041eef71221c0a39da2e9dd62ea95617