neoverse-rd/rdfremont: Add dvfs_handler driver
DVFS hardware is a hardware level abstraction to achieve frequency and
voltage scaling. DVFS hardware reduces the complexity in software to
achieve frequency and voltage scaling, by writing directly to the DVFS
hardware register and not to worry about the complex hardware logic
beneath the DVFS hardware.

The DVFS handler module is exclusively applicable to the RD-Fremont
platform, as it enables dynamic voltage and frequency scaling for
individual CPUs.

Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Change-Id: Ie9c49297b7e16d1be53b5d92e450250425b52ba4