ci: Build products in parallel in pipelines
Currently, within the deployment and daily pipeline, products are
build sequentially instead of in parallel which is slow. This patch
adds building all products in parallel within the pipelines. In the
deployment pipeline, all products are built with default log level
and in the daily pipeline all products are built with all log levels
(all in parallel).

Signed-off-by: Haaris Farooq <haaris.farooq@arm.com>